compiler + analyzers update + fixes

+++ tests/UnitTests/Common/Utility/RetryTests.cs
@@ -9,7 +9,7 @@
 
     public class RetryTests
     {
-        private static Func<uint, TimeSpan> progressive = x => TimeSpan.FromMilliseconds(
+        private static Func<int, TimeSpan> progressive = x => TimeSpan.FromMilliseconds(
             Convert.ToInt32(Math.Round((1 / (1 + Math.Exp(-x + 5))) * 100)) * 100);
 
         [Fact]
@@ -88,6 +88,7 @@
                     {
                         throw new NullReferenceException();
                     }
+
                     result = 42;
                 },
                     100.Milliseconds(),
@@ -113,7 +114,7 @@
                     throw new NullReferenceException();
                 });
             });
-            retryEx.RetryCount.ShouldBe((uint)1);
+            retryEx.RetryCount.ShouldBe(1);
             retryEx.Message.ShouldBe("retry failed after #1 attempts");
 
             counter.ShouldBe(2);
@@ -131,7 +132,7 @@
                     100.Milliseconds(),
                     100.Milliseconds());
             });
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
             counter.ShouldBe(4);
@@ -180,6 +181,7 @@
                     {
                         throw new NullReferenceException();
                     }
+
                     result = 42;
                 },
                     100.Milliseconds(),
@@ -204,7 +206,7 @@
                     throw new NullReferenceException();
                 });
             });
-            retryEx.RetryCount.ShouldBe((uint)1);
+            retryEx.RetryCount.ShouldBe(1);
             retryEx.Message.ShouldBe("retry failed after #1 attempts");
 
             counter.ShouldBe(2);
@@ -222,7 +224,7 @@
                     100.Milliseconds(),
                     100.Milliseconds());
             });
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
             counter.ShouldBe(4);
@@ -246,7 +248,7 @@
                     100.Milliseconds());
             });
 
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.InnerException.ShouldBeOfType<AggregateException>();
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
@@ -272,7 +274,7 @@
                     100.Milliseconds());
             });
 
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.InnerException.ShouldBeOfType<AggregateException>();
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
@@ -323,7 +325,7 @@
                 }, exceptionPredicate);
             });
 
-            retryEx.RetryCount.ShouldBe((uint)1);
+            retryEx.RetryCount.ShouldBe(1);
             retryEx.InnerException.ShouldBeOfType<ArgumentException>();
             retryEx.Message.ShouldBe("retry failed after #1 attempts");
 
@@ -372,7 +374,7 @@
                 return true;
             };
 
-            Func<uint, TimeSpan> delayFactory = failureCount =>
+            Func<int, TimeSpan> delayFactory = failureCount =>
             {
                 if(failureCount == 4)
                 {
@@ -395,7 +397,7 @@
                 }, exceptionPredicate, delayFactory, cts.Token);
             });
 
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.InnerException.ShouldBeOfType<ArgumentException>();
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
@@ -451,6 +453,7 @@
                     {
                         throw new NullReferenceException();
                     }
+
                     return 42;
                 });
 
@@ -468,6 +471,7 @@
                     {
                         throw new NullReferenceException();
                     }
+
                     return 42;
                 },
                     100.Milliseconds());
@@ -490,6 +494,7 @@
                     {
                         throw new NullReferenceException();
                     }
+
                     return 42;
                 },
                     100.Milliseconds(),
@@ -517,7 +522,7 @@
                     100.Milliseconds(),
                     100.Milliseconds());
             });
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
             counter.ShouldBe(4);
@@ -533,7 +538,7 @@
                     throw new NullReferenceException();
                 });
             });
-            retryEx.RetryCount.ShouldBe((uint)1);
+            retryEx.RetryCount.ShouldBe(1);
             retryEx.Message.ShouldBe("retry failed after #1 attempts");
 
             counter.ShouldBe(2);
@@ -574,6 +579,7 @@
                     {
                         throw new NullReferenceException();
                     }
+
                     return 42;
                 },
                     100.Milliseconds(),
@@ -601,7 +607,7 @@
                     100.Milliseconds(),
                     100.Milliseconds());
             });
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
             counter.ShouldBe(4);
@@ -628,7 +634,7 @@
                     100.Milliseconds());
             });
 
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.InnerException.ShouldBeOfType<AggregateException>();
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
@@ -656,7 +662,7 @@
                     100.Milliseconds());
             });
 
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.InnerException.ShouldBeOfType<AggregateException>();
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
 
@@ -714,7 +720,7 @@
                 await Retry.On(task, exceptionPredicate);
             });
 
-            retryEx.RetryCount.ShouldBe((uint)1);
+            retryEx.RetryCount.ShouldBe(1);
             retryEx.InnerException.ShouldBeOfType<ArgumentException>();
             retryEx.Message.ShouldBe("retry failed after #1 attempts");
 
@@ -767,7 +773,7 @@
                 return true;
             };
 
-            Func<uint, TimeSpan> delayFactory = failureCount =>
+            Func<int, TimeSpan> delayFactory = failureCount =>
             {
                 if(failureCount == 4)
                 {
@@ -793,7 +799,7 @@
                 }, exceptionPredicate, delayFactory, cts.Token);
             });
 
-            retryEx.RetryCount.ShouldBe((uint)3);
+            retryEx.RetryCount.ShouldBe(3);
             retryEx.InnerException.ShouldBeOfType<ArgumentException>();
             retryEx.Message.ShouldBe("retry failed after #3 attempts");
